start task, add routes and manager
