+++ Src/Core/event_type.py
@@ -8,5 +8,5 @@ from Src.Core.validator import validator
 class event_type(Enum):
     DELETE_NOMENCLATURE = 1
     CHANGE_NOMENCLATURE = 2
-    CHANGE_RANGE = 3
-    CHANGE_BLOCK_PERIOD = 4
+    SAVE_DATA_REPOSITY = 3
+    RESTORE_DATA_REPOSITY = 4
