+++ main.py
@@ -5,7 +5,7 @@ from Src.DTO.domain_prototype import domain_prototype
 from Src.Core.format_reporting import format_reporting
 from Src.Reports.report_factory import report_factory
 from Src.data_reposity import data_reposity
-from Src.settings_manager import settings_manager
+from Src.Managers.settings_manager import settings_manager
 from Src.Services.start_service import start_service
 from Src.Models.warehouse import warehouse_model
 from Src.Models.warehouse_transaction import warehouse_transaction
@@ -64,9 +64,7 @@ def put_nomenclature():
 @app.route("/api/nomenclature", methods=["PATCH"])
 def patch_nomenclature():
     request_data = request.get_json()
-    
     result = observe_service.raise_event(event_type.CHANGE_NOMENCLATURE, request_data)
-    
     return Response(result)
 
 
@@ -74,7 +72,36 @@ def patch_nomenclature():
 def delete_nomenclature(unique_code: str):
     
     result = observe_service.raise_event(event_type.DELETE_NOMENCLATURE, unique_code)
+    return Response(result)
+
+
+@app.route("/api/reports/balanse_sheet", methods=["GET"])
+def reports_balanse_sheet():
+    request_data = request.get_json()
+    date_start = request_data.get("date_start")
+    date_end = request_data.get("date_end")
+    warehouse_name = request_data.get("warehouse_name")
     
+    transactions = reposity.data[reposity.transaction_key()]
+    if not transactions:
+        return Response("Нет транзакций для пересчета.", 400)
+    
+    balanse_sheet = []
+    
+    report = report_factory(manager).create_default()
+    report.create(balanse_sheet)
+    return report.result
+
+
+@app.route("/api/reposity/save", methods=["POST"])
+def save_reposity():
+    result = observe_service.raise_event(event_type.SAVE_DATA_REPOSITY, {})
+    return Response(result)
+
+
+@app.route("/api/reposity/restore", methods=["POST"])
+def restore_reposity():
+    result = observe_service.raise_event(event_type.RESTORE_DATA_REPOSITY, {})
     return Response(result)
 
 
@@ -172,9 +199,6 @@ def reports_transaction(format: str):
 
 @app.route("/api/reports/turnovers", methods=["GET"])
 def reports_turnover(format: str):
-    inner_format = format_reporting(format)
-    report = report_factory(manager).create(inner_format)
-    
     process_turnover = turnover_process.create()
     turnovers = process_turnover.processing(reposity.data[data_reposity.transaction_key()])
